chore: merge state class into UiState class
