+++ app/src/main/java/org/sopt/official/feature/attendance/compose/AttendanceCodeDialog.kt
@@ -29,7 +29,7 @@ import org.sopt.official.designsystem.Black40
 import org.sopt.official.designsystem.Gray60
 import org.sopt.official.designsystem.SoptTheme
 import org.sopt.official.feature.attendance.compose.component.AttendanceCodeCardList
-import org.sopt.official.feature.attendance.model.AttendanceSession
+import org.sopt.official.feature.attendance.model.AttendanceUiState.Success.AttendanceDayType.AttendanceDay.MidtermAttendance.NotYet.AttendanceSession
 
 @Composable
 fun AttendanceCodeDialog(
