+++ app/src/main/java/org/sopt/official/feature/attendance/compose/AttendanceScreen.kt
@@ -39,12 +39,12 @@ import org.sopt.official.feature.attendance.compose.component.TodayAttendanceCar
 import org.sopt.official.feature.attendance.compose.component.TodayNoAttendanceCard
 import org.sopt.official.feature.attendance.compose.component.TodayNoScheduleCard
 import org.sopt.official.feature.attendance.model.AttendanceAction
-import org.sopt.official.feature.attendance.model.AttendanceDayType
-import org.sopt.official.feature.attendance.model.AttendanceHistory
-import org.sopt.official.feature.attendance.model.AttendanceResultType
 import org.sopt.official.feature.attendance.model.AttendanceUiState
-import org.sopt.official.feature.attendance.model.FinalAttendance
-import org.sopt.official.feature.attendance.model.MidtermAttendance
+import org.sopt.official.feature.attendance.model.AttendanceUiState.Success.AttendanceDayType
+import org.sopt.official.feature.attendance.model.AttendanceUiState.Success.AttendanceDayType.AttendanceDay.FinalAttendance
+import org.sopt.official.feature.attendance.model.AttendanceUiState.Success.AttendanceDayType.AttendanceDay.MidtermAttendance
+import org.sopt.official.feature.attendance.model.AttendanceUiState.Success.AttendanceHistory
+import org.sopt.official.feature.attendance.model.AttendanceUiState.Success.AttendanceResultType
 
 @Composable
 fun AttendanceScreen(state: AttendanceUiState.Success, action: AttendanceAction) {
