+++ app/src/main/java/org/sopt/official/feature/attendance/compose/component/AttendanceHistoryCard.kt
@@ -18,8 +18,8 @@ import androidx.compose.ui.unit.dp
 import kotlinx.collections.immutable.ImmutableList
 import kotlinx.collections.immutable.persistentListOf
 import org.sopt.official.designsystem.SoptTheme
-import org.sopt.official.feature.attendance.model.AttendanceHistory
-import org.sopt.official.feature.attendance.model.AttendanceResultType
+import org.sopt.official.feature.attendance.model.AttendanceUiState.Success.AttendanceHistory
+import org.sopt.official.feature.attendance.model.AttendanceUiState.Success.AttendanceResultType
 
 @Composable
 fun AttendanceHistoryCard(
