+++ app/src/main/java/org/sopt/official/feature/attendance/compose/component/FinalAttendanceCard.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.tooling.preview.PreviewParameter
 import androidx.compose.ui.tooling.preview.PreviewParameterProvider
 import androidx.compose.ui.unit.dp
 import org.sopt.official.designsystem.SoptTheme
-import org.sopt.official.feature.attendance.model.FinalAttendance
+import org.sopt.official.feature.attendance.model.AttendanceUiState.Success.AttendanceDayType.AttendanceDay.FinalAttendance
 
 @Composable
 fun FinalAttendanceCard(finalAttendance: FinalAttendance, modifier: Modifier = Modifier) {
