--- app/src/main/java/org/sopt/official/feature/attendance/model/AttendanceHistory.kt
@@ -1,7 +1,0 @@
-package org.sopt.official.feature.attendance.model
-
-data class AttendanceHistory(
-    val status: String,
-    val eventName: String,
-    val date: String,
-)
