--- app/src/main/java/org/sopt/official/feature/attendance/model/AttendanceResultType.kt
@@ -1,8 +1,0 @@
-package org.sopt.official.feature.attendance.model
-
-enum class AttendanceResultType(val type: String) {
-    ALL(type = "전체"),
-    PRESENT(type = "출석"),
-    LATE(type = "지각"),
-    ABSENT(type = "결석");
-}
