--- app/src/main/java/org/sopt/official/feature/attendance/model/AttendanceSession.kt
@@ -1,6 +1,0 @@
-package org.sopt.official.feature.attendance.model
-
-enum class AttendanceSession(val type: String) {
-    FIRST("1차 출석"),
-    SECOND("2차 출석")
-}
